fix : page change
- 기존 SignUp페이지에 있던 로직을 Login페이지로 수정완료하였습니다
( 이슈 : Login 페이지에 구현되어야할 코드를 SignUp에 구현하여, 다시 수정 작업하였습니다)

by ddoqi
