+++ screen/Main.jsx
@@ -17,10 +17,8 @@ const Main = ({ navigation: { navigate } }) => {
         <Text>Detail 이동</Text>
       </TouchableOpacity>
 
-      <TouchableOpacity
-        onPress={() => navigate('Stacks', { screen: 'SignUp' })}
-      >
-        <Text>SingUp</Text>
+      <TouchableOpacity onPress={() => navigate('Stacks', { screen: 'Login' })}>
+        <Text>Login</Text>
       </TouchableOpacity>
     </SafeAreaView>
   );
